# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MeaningBridge API — grief support, continuing bonds, journaling, assessments
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /professional/patients/{id}/engagement:
    get:
      operationId: getPatientEngagement
      tags: [professional]
      summary: Metadata-only engagement + safety signals for a patient (never any patient content).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Engagement metadata
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PatientEngagement" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AnthropicError" }
  /professional/patients/{id}/assistant:
    post:
      operationId: askProviderAssistant
      tags: [professional]
      summary: Ask the MeaningBridge provider assistant about a patient. Answers from metadata only.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProviderAssistantInput" }
      responses:
        "200":
          description: Assistant reply
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProviderAssistantReply" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AnthropicError" }
  /professional/patients/{id}/appointments:
    get:
      operationId: listPatientAppointments
      tags: [professional]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Appointments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Appointment" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AnthropicError" }
    post:
      operationId: proposeAppointment
      tags: [professional]
      summary: Propose a session time; emails the patient a confirm/decline link and mirrors to calendar when connected.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AppointmentInput" }
      responses:
        "201":
          description: Proposed appointment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AnthropicError" }
  /professional/appointments/{id}/cancel:
    post:
      operationId: cancelAppointment
      tags: [professional]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Cancelled appointment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AnthropicError" }
  /professional/calendar:
    get:
      operationId: getProviderCalendar
      tags: [professional]
      summary: The provider's calendar-sync setting and whether Google Calendar is connected.
      responses:
        "200":
          description: Calendar setting
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProviderCalendar" }
    put:
      operationId: updateProviderCalendar
      tags: [professional]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProviderCalendarInput" }
      responses:
        "200":
          description: Updated calendar setting
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProviderCalendar" }
...
components:
  schemas:
...
    PatientEngagement:
      type: object
      description: Metadata-only engagement + safety signals. Contains counts and timestamps, never any patient content.
      properties:
        patientId: { type: integer }
        status: { type: string }
        tier: { type: ["string", "null"] }
        sessionCount: { type: integer }
        lastActiveAt: { type: ["string", "null"], format: date-time }
        companionMessageCount: { type: integer }
        journalEntryCount: { type: integer }
        checkinCount: { type: integer }
        lastCheckinAt: { type: ["string", "null"], format: date-time }
        safetyEventCount: { type: integer }
        openSafetyEventCount: { type: integer }
        lastSafetyEventAt: { type: ["string", "null"], format: date-time }
      required:
        [patientId, status, sessionCount, companionMessageCount, journalEntryCount, checkinCount, safetyEventCount, openSafetyEventCount]
    ProviderAssistantInput:
      type: object
      properties:
        question: { type: string, minLength: 1 }
      required: [question]
    ProviderAssistantReply:
      type: object
      properties:
        answer: { type: string }
      required: [answer]
    Appointment:
      type: object
      properties:
        id: { type: integer }
        patientId: { type: integer }
        providerUserId: { type: integer }
        title: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        status: { type: string, enum: [proposed, confirmed, declined, cancelled] }
        location: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        googleEventId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required: [id, patientId, providerUserId, title, startsAt, endsAt, status, createdAt]
    AppointmentInput:
      type: object
      properties:
        title: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        location: { type: string }
        notes: { type: string }
      required: [startsAt, endsAt]
...
    ProviderCalendar:
      type: object
      properties:
        provider: { type: string }
        calendarId: { type: string }
        syncEnabled: { type: boolean }
        connected: { type: boolean, description: Whether a Google Calendar account is connected at the platform level. }
      required: [provider, calendarId, syncEnabled, connected]
    ProviderCalendarInput:
      type: object
      properties:
        calendarId: { type: string }
        syncEnabled: { type: boolean }
...
    AnthropicError:
      type: object
      properties:
        error: { type: string }
      required: [error]
...