# lib/api-zod/src/generated/types/patientEngagement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MeaningBridge API — grief support, continuing bonds, journaling, assessments
 * OpenAPI spec version: 0.1.0
 */

/**
 * Metadata-only engagement + safety signals. Contains counts and timestamps, never any patient content.
 */
export interface PatientEngagement {
  patientId: number;
  status: string;
  /** @nullable */
  tier?: string | null;
  sessionCount: number;
  /** @nullable */
  lastActiveAt?: Date | null;
  companionMessageCount: number;
  journalEntryCount: number;
  checkinCount: number;
  /** @nullable */
  lastCheckinAt?: Date | null;
  safetyEventCount: number;
  openSafetyEventCount: number;
  /** @nullable */
  lastSafetyEventAt?: Date | null;
}