# lib/api-zod/src/generated/types/providerCalendar.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MeaningBridge API — grief support, continuing bonds, journaling, assessments
 * OpenAPI spec version: 0.1.0
 */

export interface ProviderCalendar {
  provider: string;
  calendarId: string;
  syncEnabled: boolean;
  /** Whether a Google Calendar account is connected at the platform level. */
  connected: boolean;
}